Mac specific changes
- Added macOS specific configuration to CMakeLists.txt

- Fixed implicit casting issues while using Mac toolchain

+++ src/ast/types/generic_abstract.cpp
@@ -5,7 +5,7 @@
 
 namespace qat::ast {
 
-GenericAbstractType::GenericAbstractType(u64 _index, Identifier _name, GenericKind _kind, FileRange _range)
+GenericAbstractType::GenericAbstractType(usize _index, Identifier _name, GenericKind _kind, FileRange _range)
     : index(_index), name(std::move(_name)), kind(_kind), range(std::move(_range)) {
   ast::QatType::generics.push_back(this);
 }
